requirements.txt
google-cloud-bigquery
google-api-core
expiringdict~=1.2.1
elasticsearch~=7.8.1
invoke~=1.3.0
fabric
selenium~=3.141.0
apscheduler
Flask==2.1.0
gunicorn==20.1.0
requests~=2.24.0
webdriver-manager~=3.4.1
elasticsearch-dsl
pillow
msedge-selenium-tools
py-object-pool
expiringdict~=1.2.1
psutil
confluent-kafka
mysql-connector-python~=8.0.22
python-dateutil~=2.8.1
google-cloud-storage
markupsafe==2.0.1
